Add files via upload
The main file is "MINORPRO.C" and it uses file handling concepts and the files(text files) used here are "UPDATE1.TXT" and "VALIDUSE.TXT". This project implements Dijikstra's algorithm in Dynamic mode where the number of nodes depends on the entries done by the user and the edge length will also be provided by the user. The static part of the whole code is implemented using "Dijikstra's Algorithm" and "Ford-Fulkerson Algorithm". "VALIDUSE.TXT" maintains track of the valid users whose user-id and passwords are saved in it in a prescribed format and "UPDATE1.TXT" is used to provide entries to the model which the user has designed and it is also storing data in a prescribed format.